requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
inflection==0.5.1
mccabe==0.7.0
packaging==24.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
